fonts: switch to font awesome v5
This provides with even more icons. A few changes are made to the
default files to make it work correctly with the new version.

As a side-effect, the unnecessary white-spaces are removed from the
files that are modified while also terminating them with a newline
character correctly (this is a feature of a plugin used with the
editor, Atom).