Create PDF document from generated jekyll site
The steps to generate the file are documented.

Generation of the document requires that the responsiveness of the
page be removed failing which the document is generated like the
mobile site and seems to waste a lot of space.

The base URL of the site is changed to ensure that the relative
resources referenced in the HTML are found correctly when wkhtmltopdf
processes the document.

As wkhtmltopdf was unable to retrieve the JS hosted in the website, a
local copy of the file is included as a workaround.