generate-pdf: add note about issues with some versions of wkhtmltopdf
'wkhtmltopdf' obtained from package manager seems to have some
issues and the layout is bad. This seems to because the one
obtained from the package manager seems to be using an unpatched
version of Qt as a consequence of which the layout is not as expected.

The 'wkhtmltopdf' binary downloaded from the website seems to be
using a patched version of Qt which doesn't have this issue.